requirements.txt
Django ==4.2.16
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
requests==2.28.1
setuptools==65.5.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
wheel==0.38.4
